Fix deadlock and improve formatting
Proj4j version 1.0.0 was not thread-safe and was at the origin of
a deadlock when importing relations. Proj4j version 1.1.0 fixed this issue.

The project now uses an IntelliJ style base on Google's style to
format the source code. The codestyle.xml file can be used to configure
IntelliJ (File > Settings > Code Style > Import Scheme).wq