Add batch parameters and improve config accordingly
The batch parameters enables baremap to integrate with AWS Batch (#62). As a result, the generation of tiles can be parallelized and distributed.

As the extent of the map is needed to list its tiles, the configuration files now contain bounding information. This is preferable than using the OSM headers because baremaps is supposed to work with different datasets.