Fix performance issue on PG Tile Store (#909, #912)
The query optimizer of Postgres 15 is not able to push conditions down when using sub-queries. Therefore, we now append the spatial conditions to the sub-query. The unit tests have been adapted accordingly. A more robust solution involving a query parser will be devised (#913) in the future to account for more complex queries (GROUP BY, HAVING, etc.).

---------

Co-authored-by: AlexGacon <alexandre.gacon@gmail.com>